requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.2
python-dotenv
SQLAlchemy
Werkzeug
email-validator
gunicorn
colorama==0.4.6